api/pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "telexkcdbot-api"
version = "1.0.0"
description = "telexkcdbot API microservice"
authors = ["coryphoenixxx <coryphoenixxx@gmail.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/coryphoenixxx/telexkcdbot/api"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.108"
python-multipart = "^0.0.6"
orjson = "^3.9.10"
sqlalchemy = "^2.0.23"
asyncpg = "^0.29.0"
uvicorn = { extras = ["standart"], version = "^0.25" }
pydantic-settings = "^2.1.0"
aiofiles = "^23.2.1"
alembic = "^1.13.0"
pillow = "^10.1.0"
uvloop = "^0.19.0"
filetype = "^1.2.0"
imagesize = "^1.4.1"
python-slugify = "^8.0.1"
httptools = "^0.6.1"

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = ["-s", "-v"]
pythonopath = [".venv/bin"]
